lang

+++ app/Http/Controllers/BlogCategoryController.php
@@ -6,10 +6,22 @@ use App\Models\BlogCategory;
 use App\Http\Requests\StoreBlogCategoryRequest;
 use App\Http\Requests\UpdateBlogCategoryRequest;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
-class BlogCategoryController extends Controller
+class BlogCategoryController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            // new Middleware('auth', only: ['store']),
+            // new Middleware('auth', except: ['index', 'store']),
+            new Middleware('auth'),
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -38,7 +50,7 @@ class BlogCategoryController extends Controller
 
         $slug = Str::slug($fields['name']);
 
-        BlogCategory::create([
+        Auth::user()->blogCategories()->create([
             'name' => $fields['name'],
             'slug' => $slug
         ]);
@@ -68,6 +80,8 @@ class BlogCategoryController extends Controller
      */
     public function update(Request $request, BlogCategory $blogCategory)
     {
+        Gate::authorize('modify', $blogCategory);
+
         $fields = $request->validate([
             'name' => 'required|string|max:255'
         ]);
@@ -87,6 +101,8 @@ class BlogCategoryController extends Controller
      */
     public function destroy(BlogCategory $blogCategory)
     {
+        Gate::authorize('modify', $blogCategory);
+
         $blogCategory->delete();
 
         return back()->with('delete', 'Blog Category deleted successfully');
